Move the getAllResources method to a trait for reuse
